+++ packages/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @a-little-world/little-world-design-system
 
+## 2.15.1
+
+### Patch Changes
+
+- Fix Button Cirlce instances
+- Add prop closeOnBackdropClick to Modal
+
 ## 2.15.0
 
 ### Minor Changes